Add NumPy 2 support (#605)

+++ tests/test_stats.py
@@ -18,7 +18,7 @@ class TestStats:
 
         # Check that the NMAD is computed the same with a masked array or NaN array, and is equal to scipy nmad
         nmad_ma = nmad(self.landsat_raster.data)
-        nmad_array = nmad(self.landsat_raster.get_nanarray())
+        nmad_array = nmad(self.landsat_raster.get_nanarray(floating_dtype="float64"))
         nmad_scipy = scipy.stats.median_abs_deviation(self.landsat_raster.data, axis=None, scale="normal")
 
         assert nmad_ma == nmad_array
